Move EntityExcelSheetHandler to it's own class to make it more testable
